Добавил filmPresenter
