+++ src/main.js
@@ -1,6 +1,6 @@
 import SiteMenuView from './view/site-menu-view.js';
 import HeaderProfileView from './view/header-profile-view.js';
-import MovieListPresenter from './presenter/card-presenter.js';
+import FilmListPresenter from './presenter/film-list-presenter.js';
 import { render, RenderPosition } from './utils/render.js';
 import { generateCard } from './mock/card-movie.js';
 import { generateFilter } from './mock/filter.js';
@@ -11,7 +11,7 @@ const filters = generateFilter(cards);
 const siteMainElement = document.querySelector('.main');
 const siteNavigationElement = document.querySelector('.header');
 
-const cardPresenter = new MovieListPresenter(siteMainElement);
+const filmListPresenter = new FilmListPresenter(siteMainElement);
 render(siteMainElement, new SiteMenuView(filters), RenderPosition.BEFOREEND);
 render(siteNavigationElement, new HeaderProfileView(), RenderPosition.BEFOREEND);
-cardPresenter.init(cards);
+filmListPresenter.init(cards);
